+++ src/Cli/func/Actions/HostActions/StartHostAction.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System.Collections;
@@ -460,7 +460,24 @@ namespace Azure.Functions.Cli.Actions.HostActions
 
             (var listenUri, var baseUri, var certificate) = await Setup();
 
-            await ExtensionBundleHelper.GetExtensionBundle();
+            // Determine if we should download extension bundles
+            // Priority: 1) Environment variable, 2) host.json, 3) Default (download enabled)
+            var (shouldDownloadBundles, source) = ShouldDownloadExtensionBundles();
+
+            if (shouldDownloadBundles)
+            {
+                if (isVerbose)
+                {
+                    ColoredConsole.WriteLine(VerboseColor("Downloading extension bundles..."));
+                }
+
+                await ExtensionBundleHelper.GetExtensionBundle();
+            }
+            else if (isVerbose)
+            {
+                ColoredConsole.WriteLine(VerboseColor($"Skipping extension bundle download because ensureLatest is set to false in {source}."));
+            }
+
             IWebHost host = await BuildWebHost(hostOptions, listenUri, baseUri, certificate);
             var runTask = host.RunAsync();
             var hostService = host.Services.GetRequiredService<WebJobsScriptHostService>();
@@ -891,6 +908,39 @@ namespace Azure.Functions.Cli.Actions.HostActions
         private void PrintMigrationWarningForDotnet6Inproc()
         {
             ColoredConsole.WriteLine(WarningColor($".NET 6 is no longer supported. Please consider migrating to a supported version. For more information, see https://aka.ms/azure-functions/dotnet/net8-in-process. If you intend to target .NET 8 on the in-process model, make sure that '{Constants.InProcDotNet8EnabledSetting}' is set to '1' in {Constants.LocalSettingsJsonFileName}.\n"));
+        }
+
+        /// <summary>
+        /// Determines whether extension bundles should be downloaded.
+        /// Priority: 1) Environment variable, 2) host.json, 3) Default (download enabled).
+        /// </summary>
+        /// <returns>A tuple indicating whether to download bundles and the source of the setting.</returns>
+        internal (bool ShouldDownload, string Source) ShouldDownloadExtensionBundles()
+        {
+            const string envVarKey = "AzureFunctionsJobHost__extensionBundle__ensureLatest";
+            const string hostJsonConfigKey = "AzureFunctionsJobHost:extensionBundle:ensureLatest";
+
+            // 1. Check environment variable first (highest priority)
+            var ensureLatestEnvVar = Environment.GetEnvironmentVariable(envVarKey);
+            if (bool.TryParse(ensureLatestEnvVar, out var ensureLatestFromEnv))
+            {
+                // ensureLatest=false means we should NOT download (skip to use cached)
+                // ensureLatest=true means we SHOULD download
+                return (ensureLatestFromEnv, "environment variable");
+            }
+
+            // 2. Check host.json configuration (secondary option)
+            if (_hostJsonConfig != null)
+            {
+                var ensureLatestFromHostJson = _hostJsonConfig[hostJsonConfigKey];
+                if (bool.TryParse(ensureLatestFromHostJson, out var ensureLatestFromConfig))
+                {
+                    return (ensureLatestFromConfig, "host.json");
+                }
+            }
+
+            // 3. Default behavior: download bundles (shouldDownload = true)
+            return (true, string.Empty);
         }
 
         /// <summary>